Core/PacketIO : Updated Some Guild and GuildFinder Packets
+ Updated Following packets :

CMSG_GUILD_BANK_MONEY_WITHDRAWN_QUERY , CMSG_GUILD_BANK_WITHDRAW_MONEY
CMSG_GUILD_SET_NOTE , CMSG_LF_GUILD_BROWSE , CMSG_LF_GUILD_GET_APPLICATIONS
SMSG_GUILD_BANK_MONEY_WITHDRAWN , SMSG_LF_GUILD_BROWSE_UPDATED
SMSG_LF_GUILD_MEMBERSHIP_LIST_UPDATED

+++ src/server/game/Guilds/Guild.cpp
@@ -2329,7 +2329,7 @@ void Guild::SendMoneyInfo(WorldSession* session) const
     if (!member)
         return;
 
-    int32 amount = _GetMemberRemainingMoney(member);
+    int64 amount = _GetMemberRemainingMoney(member);
     WorldPacket data(SMSG_GUILD_BANK_MONEY_WITHDRAWN, 8);
     data << int64(amount);
     session->SendPacket(&data);
